Add checkpoints to the end of each section.
**All sections**

Jason Fehr also pointed out that there are no check points where you can
stop and view the progress of your app to make sure it is functioning
correctly. A checkpoint has been added to the end of each section for
this purpose. These checkpoints are clearly marked by a “Checkpoint”
sub-heading.